+++ packages/popper/src/index.ts
@@ -1,3 +1,4 @@
-export * from './popper.modifiers';
-export * from './popper.utils';
-export * from './use-popper';
+export type { Placement, PlacementWithLogical } from './popper.placement';
+export { cssVars as popperCSSVars } from './popper.utils';
+export { usePopper } from './use-popper';
+export type { UsePopperProps, UsePopperReturn } from './use-popper';