+++ packages/system/src/jsx.tsx
@@ -1,4 +1,4 @@
-import { Interpolation, jsx as emotion, Theme } from '@emotion/react';
+import { Interpolation, Theme, jsx as emotion } from '@emotion/react';
 import { Dict } from '@nature-ui/utils';
 import React from 'react';