+++ scripts/setup-test.ts
@@ -1,3 +1,3 @@
-//@ts-expect-error
+import '@testing-library/jest-dom/extend-expect';
 const { getComputedStyle } = window;
 window.getComputedStyle = (elt) => getComputedStyle(elt);