+++ packages/input/CHANGELOG.md
@@ -1,5 +1,172 @@
 # @nature-ui/input
 
+## 2.2.13
+
+### Patch Changes
+
+- Fixing release bug on @nature-ui/toast and portal.
+- Updated dependencies
+  - @nature-ui/form-control@2.2.13
+
+## 2.2.12
+
+### Patch Changes
+
+- Fix error on @nature-ui/image component
+- Updated dependencies
+  - @nature-ui/form-control@2.2.12
+
+## 2.2.11
+
+### Patch Changes
+
+- Fix previous release
+- Updated dependencies
+  - @nature-ui/form-control@2.2.11
+
+## 0.0.1
+
+### Patch Changes
+
+- Export module
+- Updated dependencies
+  - @nature-ui/form-control@0.0.1
+
+## 0.0.0
+
+### Patch Changes
+
+- Popover fix
+- Updated dependencies
+  - @nature-ui/form-control@0.0.0
+
+## 0.0.0-dev-20230411071731
+
+### Patch Changes
+
+-
+- Updated dependencies
+  - @nature-ui/form-control@0.0.0-dev-20230411071731
+
+## 0.0.0-dev-20230410162901
+
+### Patch Changes
+
+- popover
+- Updated dependencies
+  - @nature-ui/form-control@0.0.0-dev-20230410162901
+
+## 0.0.0-dev-20230410131720
+
+### Patch Changes
+
+- fix tailwind error
+- Updated dependencies
+  - @nature-ui/form-control@0.0.0-dev-20230410131720
+
+## 0.0.0-dev-20230409203337
+
+### Patch Changes
+
+- prepack
+- Updated dependencies
+  - @nature-ui/form-control@0.0.0-dev-20230409203337
+
+## 0.0.0-dev-20230409203113
+
+### Patch Changes
+
+- fix mjs build
+- Updated dependencies
+  - @nature-ui/form-control@0.0.0-dev-20230409203113
+
+## 0.0.0-dev-20230409201910
+
+### Patch Changes
+
+- configure prepack
+- Updated dependencies
+  - @nature-ui/form-control@0.0.0-dev-20230409201910
+
+## 2.2.1
+
+### Patch Changes
+
+- Add popover component
+- Updated dependencies
+  - @nature-ui/form-control@2.2.1
+
+## 2.2.0
+
+### Minor Changes
+
+- v2
+
+### Patch Changes
+
+- Updated dependencies
+  - @nature-ui/form-control@2.2.0
+
+## 2.1.0
+
+### Minor Changes
+
+- v2
+
+### Patch Changes
+
+- Updated dependencies
+  - @nature-ui/form-control@2.1.0
+
+## 2.0.0
+
+### Major Changes
+
+- v2
+
+### Patch Changes
+
+- Updated dependencies
+  - @nature-ui/form-control@2.0.0
+
+## 1.0.0
+
+### Major Changes
+
+- v2
+
+### Patch Changes
+
+- Updated dependencies
+  - @nature-ui/form-control@1.0.0
+
+## 0.0.0
+
+### Major Changes
+
+- Releasing v2
+
+### Patch Changes
+
+- Updated dependencies
+  - @nature-ui/form-control@0.0.0
+
+## 0.0.0-dev-20230301133840
+
+### Patch Changes
+
+- update react version
+- Updated dependencies
+  - @nature-ui/form-control@0.0.0-dev-20230301133840
+
+## 0.0.0-dev-20230301120121
+
+### Patch Changes
+
+- dev
+- Updated dependencies
+  - @nature-ui/form-control@0.0.0-dev-20230301120121
+
 ## 0.0.0-dev-20221217102739
 
 ### Patch Changes